[Dheer] Add: Implemented the functionality of Insert new contact to t…

+++ AddressBookProblem-ADO.NET/AddressBookModel.cs
@@ -20,7 +20,7 @@ namespace AddressBookProblem_ADO.NET
         public string Address { get; set; }
         public string City { get; set; }
         public string State { get; set; }
-        public long Zip { get; set; }
+        public int Zip { get; set; }
         public long PhoneNumber { get; set; }
         public string EmailId { get; set; }
         public string AddressBookType { get; set; }
